requirements.txt
opencv-python
Pillow
sklearn
numpy
tk